Remove surf specific code. (#7)
* Removed unused config

* Removed SURF specific code from the orchestrator-core

* Moved network type validators to separate file

* Bumpversion
