+++ .bumpversion.cfg
@@ -1,9 +1,9 @@
 [bumpversion]
-current_version = 0.0.10
+current_version = 0.0.11
 commit = False
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(\-(?P<release>[a-z]+)(?P<build>\d+))?
-serialize = 
+serialize =
 	{major}.{minor}.{patch}-{release}{build}
 	{major}.{minor}.{patch}
