+++ orchestrator/config/__init__.py
@@ -10,33 +10,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-
-# Fixed inputs
-CITY_TYPE = "city_type"
-DOMAIN = "domain"
-PORT_SPEED = "port_speed"
-PROTECTION_TYPE = "protection_type"
-SERVICE_SPEED = "service_speed"
-AGGREGATE = "aggregate"
-REDUNDANT = "redundant"
-TAGGED = "tagged"
-NETHERLIGHT = "netherlight"
-IP_ROUTING_TYPE = "ip_routing_type"
-SIZE = "size"
-
-# Resource types
-PORT_SUBSCRIPTION_ID = "port_subscription_id"
-IMS_CIRCUIT_ID = "ims_circuit_id"
-IMS_NODE_ID = "ims_node_id"
-IPAM_PREFIX_ID = "ipam_prefix_id"
-IP_PREFIX_SUBSCRIPTION_ID = "ip_prefix_subscription_id"
-PARENT_IP_PREFIX_SUBSCRIPTION_ID = "parent_ip_prefix_subscription_id"
-NSO_DEVICE_ID = "nso_device_id"  # will contain the NSO device name
-NSO_SERVICE_ID = "nso_service_id"  # will contain the service UUID
-NODE_SUBSCRIPTION_ID = "node_subscription_id"
-PORT_MODE = "port_mode"
-INTERNETPINNEN_PREFIX_SUBSCRIPTION_ID = "internetpinnen_prefix_subscription_id"
-PEER_GROUP_SUBSCRIPTION_ID = "peer_group_subscription_id"
-
-MAX_SPEED_POSSIBLE = 1000_000
