+++ test/unit_tests/config.py
@@ -1,0 +1,11 @@
+CITY_TYPE = "city_type"
+DOMAIN = "domain"
+PORT_SPEED = "port_speed"
+SERVICE_SPEED = "service_speed"
+PORT_SUBSCRIPTION_ID = "port_subscription_id"
+IMS_CIRCUIT_ID = "ims_circuit_id"
+IPAM_PREFIX_ID = "ipam_prefix_id"
+PARENT_IP_PREFIX_SUBSCRIPTION_ID = "parent_ip_prefix_subscription_id"
+NODE_SUBSCRIPTION_ID = "node_subscription_id"
+INTERNETPINNEN_PREFIX_SUBSCRIPTION_ID = "internetpinnen_prefix_subscription_id"
+PEER_GROUP_SUBSCRIPTION_ID = "peer_group_subscription_id"
